\name{exmask}
\docType{data}
\alias{exmask}
\title{Output object of the function mask}
\description{This data is the output object of the function \code{mask}
  for the AffyBatch object \code{newAffyBatch}.}
\usage{exmask}
\format{List of 1 or 2 objects.}
\source{??}
\references{??}
\keyword{datasets}