\name{newAffyBatch}
\docType{data}
\alias{newAffyBatch}
\title{AffyBatch with reduced genes}
\description{This data is an AffyBatch object with a subset of 100 genes
  with human chimpanzee data (cdf hgu95av2) - 10 individuals each.}
\usage{newAffyBatch}
\format{AffyBatch object}
\source{??}
\references{Khaitovich et al.,Parallel Patterns of Evolution in the
  Genomes and Transcriptomes of Humans and Chimpanzees, Science 2005}
\keyword{datasets}