\name{sequenceMask}
\docType{data}
\alias{sequenceMask}
\title{Object containing sequence information for probes.}
\description{This data is a table with information about sequence
  difference between human and chimpanzee for all available probes.}
\usage{sequenceMask}
\format{data.frame.}
\source{??}
\references{??}
\keyword{datasets}