\name{sequenceMask}
\docType{data}
\alias{sequenceMask}
\title{Output object of the function seq.mask}
\description{This data is a table with information about sequence
  difference between human and chimpanzee for all available probes.}
\usage{sequenceMask}
\format{List of 2 objects.}
\source{??}
\references{??}
\keyword{datasets}